Add files via upload
Parkinson disease prediction system uses Numpy which is  Python library used for working with arrays. It also has functions for working in domain of linear algebra, fourier transform, and matrices. The system also uses panda which is a  fast, powerful, flexible and easy to use open source data analysis and manipulation tool, built on top of the Python programming language. The model utilizes scikit library ,the most useful and robust library for machine learning in Python. It provides a selection of efficient tools for machine learning and statistical modeling including classification, regression, clustering and dimensionality reduction via a consistence interface in Python.